+++ pruebas-de-concepto/app-backend/backend/src/app.js
@@ -1,0 +1,12 @@
+const express = require("express");
+const logger = require("morgan");
+
+const indexRouter = require("./routes/index");
+
+const app = express();
+
+app.use(logger("dev"));
+
+app.use("/", indexRouter);
+
+module.exports = app;
