--- pruebas-de-concepto/app-backend/src/app.js
@@ -1,18 +1,0 @@
-const express = require("express");
-const cookieParser = require("cookie-parser");
-const logger = require("morgan");
-
-const indexRouter = require("./routes/index");
-const usersRouter = require("./routes/users");
-
-const app = express();
-
-app.use(logger("dev"));
-app.use(express.json());
-app.use(express.urlencoded({ extended: false }));
-app.use(cookieParser());
-
-app.use("/", indexRouter);
-app.use("/users", usersRouter);
-
-module.exports = app;
