--- pruebas-de-concepto/app-backend/src/routes/index.js
@@ -1,10 +1,0 @@
-const express = require("express");
-
-const router = express.Router();
-
-/* GET home page. */
-router.get("/", (req, res) => {
-  res.send("Hello World!");
-});
-
-module.exports = router;
