--- pruebas-de-concepto/app-backend/src/routes/users.js
@@ -1,10 +1,0 @@
-const express = require("express");
-
-const router = express.Router();
-
-/* GET users listing. */
-router.get("/", (req, res) => {
-  res.send("respond with a resource");
-});
-
-module.exports = router;
