+++ pruebas-de-concepto/bt_bidireccional/arduino/arduino.ino.standard.hex
@@ -1,0 +1,119 @@
+:100000000C9435000C945D000C945D000C945D0024
+:100010000C945D000C945D000C945D000C945D00EC
+:100020000C945D000C945D000C945D000C945D00DC
+:100030000C945D000C945D000C945D000C945D00CC
+:100040000C94E9010C945D000C94B7010C9491019F
+:100050000C945D000C945D000C945D000C945D00AC
+:100060000C945D000C945D00480311241FBECFEF7B
+:10007000D8E0DEBFCDBF11E0A0E0B1E0EEE3F7E0F5
+:1000800002C005900D92A631B107D9F721E0A6E193
+:10009000B1E001C01D92AE3BB207E1F710E0C5E34D
+:1000A000D0E004C02197FE010E949703C433D1071A
+:1000B000C9F70E9433020C949D030C940000AF9288
+:1000C000BF92CF92DF92EF92FF920F931F93CF9345
+:1000D000DF936C017B018B01040F151FEB015E01A7
+:1000E000AE18BF08C017D10759F06991D601ED913C
+:1000F000FC910190F081E02DC6010995892B79F7DB
+:10010000C501DF91CF911F910F91FF90EF90DF908C
+:10011000CF90BF90AF900895FC01538D448D252F53
+:1001200030E0842F90E0821B930B541710F0CF9691
+:10013000089501970895FC01918D828D981761F0C3
+:10014000A28DAE0FBF2FB11D5D968C91928D9F5FDA
+:100150009F73928F90E008958FEF9FEF0895FC01B9
+:10016000918D828D981731F0828DE80FF11D858D6C
+:1001700090E008958FEF9FEF0895FC01918D228DFF
+:10018000892F90E0805C9F4F821B91098F73992784
+:1001900008958FE191E00E94BD0021E0892B09F4D0
+:1001A00020E0822F089580E090E0892B29F00E94C2
+:1001B000C90081110C9400000895FC01A48DA80FC2
+:1001C000B92FB11DA35ABF4F2C91848D90E0019699
+:1001D0008F739927848FA689B7892C93A089B189B9
+:1001E0008C91837080648C93938D848D981306C05A
+:1001F0000288F389E02D80818F7D80830895EF92BE
+:10020000FF920F931F93CF93DF93EC0181E0888FD0
+:100210009B8D8C8D98131AC0E889F989808185FFA0
+:1002200015C09FB7F894EE89FF896083E889F98942
+:1002300080818370806480839FBF81E090E0DF9144
+:10024000CF911F910F91FF90EF900895F62E0B8D97
+:1002500010E00F5F1F4F0F731127E02E8C8D8E1152
+:100260000CC00FB607FCFACFE889F989808185FFB9
+:10027000F5CFCE010E94DD00F1CFEB8DEC0FFD2F0D
+:10028000F11DE35AFF4FF0829FB7F8940B8FEA8974
+:10029000FB8980818062CFCFCF93DF93EC01888D83
+:1002A0008823B9F0AA89BB89E889F9898C9185FDF1
+:1002B00003C0808186FD0DC00FB607FCF7CF8C917F
+:1002C00085FFF2CF808185FFEDCFCE010E94DD005A
+:1002D000E9CFDF91CF9108953FB7F89480911B014A
+:1002E00090911C01A0911D01B0911E0126B5A89B03
+:1002F00005C02F3F19F00196A11DB11D3FBFBA2FB8
+:10030000A92F982F8827BC01CD01620F711D811D77
+:10031000911D42E0660F771F881F991F4A95D1F7FC
+:1003200008951F920F920FB60F9211242F933F93AF
+:100330004F935F936F937F938F939F93AF93BF93ED
+:10034000EF93FF938FE191E00E94DD00FF91EF9129
+:10035000BF91AF919F918F917F916F915F914F91DD
+:100360003F912F910F900FBE0F901F9018951F92E5
+:100370000F920FB60F9211242F938F939F93EF93A9
+:10038000FF93E0912F01F09130018081E0913501E0
+:10039000F091360182FD1BC09081809138018F5F02
+:1003A0008F7320913901821741F0E0913801F0E01C
+:1003B000E15EFE4F958F80933801FF91EF919F9101
+:1003C0008F912F910F900FBE0F901F9018958081E5
+:1003D000F4CF1F920F920FB60F9211242F933F93D9
+:1003E0008F939F93AF93BF938091170190911801C2
+:1003F000A0911901B0911A013091160123E0230F49
+:100400002D3758F50196A11DB11D2093160180933B
+:10041000170190931801A0931901B0931A018091CC
+:100420001B0190911C01A0911D01B0911E0101962C
+:10043000A11DB11D80931B0190931C01A0931D0170
+:10044000B0931E01BF91AF919F918F913F912F91DA
+:100450000F900FBE0F901F90189526E8230F02965D
+:10046000A11DB11DD2CFCF93DF93CDB7DEB7A1973A
+:100470000FB6F894DEBF0FBECDBF789484B582600E
+:1004800084BD84B5816084BD85B5826085BD85B538
+:10049000816085BD80916E00816080936E001092B6
+:1004A0008100809181008260809381008091810031
+:1004B00081608093810080918000816080938000C2
+:1004C0008091B10084608093B1008091B000816020
+:1004D0008093B00080917A00846080937A0080914C
+:1004E0007A00826080937A0080917A0081608093A4
+:1004F0007A0080917A00806880937A001092C1001F
+:10050000E0912F01F091300182E08083E0912B0196
+:10051000F0912C011082E0912D01F0912E018FECD1
+:10052000808310923701E0913301F091340186E02D
+:100530008083E0913101F09132018081806180837C
+:10054000E0913101F0913201808188608083E091F7
+:100550003101F0913201808180688083E091310126
+:10056000F091320180818F7D808380E4782E9AE043
+:10057000892E912CA12CB12C00E010E070927C000F
+:1005800080917A00806480937A0080917A0086FD61
+:10059000FCCF20917800C0907900DC2CC22E809195
+:1005A000BC019091BD018C159D0509F440C0D0920D
+:1005B000BD01C092BC010D2C000CEE08FF08F7FE37
+:1005C00051C06DE28FE191E00E94FF0022273327A6
+:1005D000A9012C193D094E095F0919A2CE01819686
+:1005E0007C01CA01B901A50194010E947503605DF7
+:1005F000F70162937F01211531054105510589F706
+:10060000E114F10461F001900020E9F73197AF01A6
+:100610004E195F09B7018FE191E00E945F0042E04F
+:1006200050E062E171E08FE191E00E945F000E9482
+:100630006C016B017C010E946C016C197D098E09B3
+:100640009F09683E734081059105A8F301151105C6
+:1006500009F494CF0E94C900882309F48FCF0E9427
+:1006600000008CCFA701960119A2CE0181967C01D2
+:10067000CA01B901A50194010E947503605DF701EB
+:1006800062937F01211531054105510589F7B8CFE6
+:10069000EFE1F1E01382128288EE93E0A0E0B0E097
+:1006A00084839583A683B78384E091E091838083DC
+:1006B00085EC90E09587848784EC90E09787868727
+:1006C00080EC90E0918B808B81EC90E0938B828B1F
+:1006D00082EC90E0958B848B86EC90E0978B868BF8
+:1006E000118E128E138E148E0895A1E21A2EAA1B5B
+:1006F000BB1BFD010DC0AA1FBB1FEE1FFF1FA217D2
+:10070000B307E407F50720F0A21BB30BE40BF50BCE
+:10071000661F771F881F991F1A9469F76095709557
+:10072000809590959B01AC01BD01CF010895EE0F1E
+:0E073000FF1F0590F491E02D0994F894FFCF7F
+:10073E0000000000FF005F008C004C01BD009B001C
+:06074E00AF000D0A0000DF
+:00000001FF
