+++ pruebas-de-concepto/bt_unidireccional/arduino/arduino.ino.standard.hex
@@ -1,0 +1,118 @@
+:100000000C945D000C9485000C9485000C94850084
+:100010000C9485000C9485000C9485000C9485004C
+:100020000C9485000C9485000C9485000C9485003C
+:100030000C9485000C9485000C9485000C9485002C
+:100040000C9459020C9485000C9427020C94010224
+:100050000C9485000C9485000C9485000C9485000C
+:100060000C9485000C9485000000000024002700FB
+:100070002A0000000000250028002B0004040404CE
+:100080000404040402020202020203030303030342
+:10009000010204081020408001020408102001021F
+:1000A00004081020000000080002010000030407FB
+:1000B0000000000000000000630311241FBECFEF0A
+:1000C000D8E0DEBFCDBF11E0A0E0B1E0E0E3F7E0B3
+:1000D00002C005900D92A231B107D9F721E0A2E14B
+:1000E000B1E001C01D92A83BB207E1F710E0CDE5F9
+:1000F000D0E004C02197FE010E949003CC35D107C7
+:10010000C9F70E94A3020C9496030C940000E0EB44
+:10011000F0E02491ECE9F0E09491E8E8F0E0E4917B
+:10012000EE23C9F0222339F0233001F1A8F4213065
+:1001300019F1223029F1F0E0EE0FFF1FEE58FF4FCA
+:10014000A591B4912FB7F894EC91811126C09095A8
+:100150009E239C932FBF08952730A9F02830C9F023
+:10016000243049F7209180002F7D03C0209180002A
+:100170002F7720938000DFCF24B52F7724BDDBCFEE
+:1001800024B52F7DFBCF2091B0002F772093B000B6
+:10019000D2CF2091B0002F7DF9CF9E2BDACFAF9236
+:1001A000BF92CF92DF92EF92FF920F931F93CF9364
+:1001B000DF936C017B018B01040F151FEB015E01C6
+:1001C000AE18BF08C017D10759F06991D601ED915B
+:1001D000FC910190F081E02DC6010995892B79F7FA
+:1001E000C501DF91CF911F910F91FF90EF90DF90AC
+:1001F000CF90BF90AF900895FC01538D448D252F73
+:1002000030E0842F90E0821B930B541710F0CF96B0
+:10021000089501970895FC01918D828D981761F0E2
+:10022000A28DAE0FBF2FB11D5D968C91928D9F5FF9
+:100230009F73928F90E008958FEF9FEF0895FC01D8
+:10024000918D828D981731F0828DE80FF11D858D8B
+:1002500090E008958FEF9FEF0895FC01918D228D1E
+:10026000892F90E0805C9F4F821B91098F739927A3
+:1002700008958BE191E00E942D0121E0892B09F482
+:1002800020E0822F089580E090E0892B29F00E94E1
+:10029000390181110C9400000895FC01A48DA80F70
+:1002A000B92FB11DA35ABF4F2C91848D90E00196B8
+:1002B0008F739927848FA689B7892C93A089B189D8
+:1002C0008C91837080648C93938D848D981306C079
+:1002D0000288F389E02D80818F7D80830895EF92DD
+:1002E000FF920F931F93CF93DF93EC0181E0888FF0
+:1002F0009B8D8C8D98131AC0E889F989808185FFC0
+:1003000015C09FB7F894EE89FF896083E889F98961
+:1003100080818370806480839FBF81E090E0DF9163
+:10032000CF911F910F91FF90EF900895F62E0B8DB6
+:1003300010E00F5F1F4F0F731127E02E8C8D8E1171
+:100340000CC00FB607FCFACFE889F989808185FFD8
+:10035000F5CFCE010E944D01F1CFEB8DEC0FFD2FBB
+:10036000F11DE35AFF4FF0829FB7F8940B8FEA8993
+:10037000FB8980818062CFCFCF93DF93EC01888DA2
+:100380008823B9F0AA89BB89E889F9898C9185FD10
+:1003900003C0808186FD0DC00FB607FCF7CF8C919E
+:1003A00085FFF2CF808185FFEDCFCE010E944D0108
+:1003B000E9CFDF91CF9108953FB7F894809117016D
+:1003C00090911801A0911901B0911A0126B5A89B2E
+:1003D00005C02F3F19F00196A11DB11D3FBFBA2FD7
+:1003E000A92F982F8827BC01CD01620F711D811D97
+:1003F000911D42E0660F771F881F991F4A95D1F71C
+:1004000008951F920F920FB60F9211242F933F93CE
+:100410004F935F936F937F938F939F93AF93BF930C
+:10042000EF93FF938BE191E00E944D01FF91EF91DB
+:10043000BF91AF919F918F917F916F915F914F91FC
+:100440003F912F910F900FBE0F901F9018951F9204
+:100450000F920FB60F9211242F938F939F93EF93C8
+:10046000FF93E0912B01F0912C018081E09131010B
+:10047000F091320182FD1BC09081809134018F5F29
+:100480008F7320913501821741F0E0913401F0E043
+:10049000E55EFE4F958F80933401FF91EF919F9120
+:1004A0008F912F910F900FBE0F901F901895808104
+:1004B000F4CF1F920F920FB60F9211242F933F93F8
+:1004C0008F939F93AF93BF938091130190911401E9
+:1004D000A0911501B09116013091120123E0230F74
+:1004E0002D3758F50196A11DB11D2093120180935F
+:1004F000130190931401A0931501B09316018091FC
+:10050000170190911801A0911901B0911A0101965B
+:10051000A11DB11D8093170190931801A09319019B
+:10052000B0931A01BF91AF919F918F913F912F91FD
+:100530000F900FBE0F901F90189526E8230F02967C
+:10054000A11DB11DD2CF789484B5826084BD84B5DD
+:10055000816084BD85B5826085BD85B5816085BDBE
+:1005600080916E00816080936E0010928100809176
+:1005700081008260809381008091810081608093FE
+:100580008100809180008160809380008091B10023
+:1005900084608093B1008091B00081608093B0004E
+:1005A00080917A00846080937A0080917A008260E2
+:1005B00080937A0080917A00816080937A008091A4
+:1005C0007A00806880937A001092C100ECE9F0E034
+:1005D0002491E8E8F0E08491882399F090E0880F76
+:1005E000991FFC01E859FF4FA591B491FC01EE5809
+:1005F000FF4F859194918FB7F894EC91E22BEC9397
+:100600008FBFE0912B01F0912C0182E08083E0917B
+:100610002701F09128011082E0912901F0912A012F
+:100620008FEC808310923301E0912F01F091300123
+:1006300086E08083E0912D01F0912E018081806120
+:100640008083E0912D01F0912E018081886080836C
+:10065000E0912D01F0912E01808180688083E091EE
+:100660002D01F0912E0180818F7D8083C0E0D0E04C
+:100670008BE191E00E942D01892B49F08BE191E003
+:100680000E940B018136D9F481E00E9487000E940C
+:10069000DC016B017C010E94DC016C197D098E0973
+:1006A0009F09683E734081059105A8F3209701F3E7
+:1006B0000E9439018823E1F20E940000D9CF8236DE
+:1006C00031F780E0E2CFEBE1F1E01382128288EEB5
+:1006D00093E0A0E0B0E084839583A683B78384E0B1
+:1006E00091E09183808385EC90E09587848784EC0A
+:1006F00090E09787868780EC90E0918B808B81ECEF
+:1007000090E0938B828B82EC90E0958B848B86ECCF
+:1007100090E0978B868B118E128E138E148E089517
+:10072000EE0FFF1F0590F491E02D0994F894FFCF90
+:10073000000000006F01CF00FC00BC012D010B0187
+:020740001F0197
+:00000001FF
